requirements.txt
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
schedule==1.1.0
